constantsを削除してhooksを更新
