--- src/constants.ts
@@ -1,12 +1,0 @@
-export const event = [
-    { value: 'event1', label: 'event1' },
-    { value: 'event2', label: 'event2' },
-    { value: 'event3', label: 'event3' },
-]
-
-export const period = [
-    {value: "1", label: "days"},
-    {value: "2", label: "weeks"},
-    {value: "3", label: "months"},
-    {value: "4", label: "years"},
-]
